requirements.txt
bs4
pandas
selenium
requests
python-dotenv
apscheduler 
openpyxl
pymongo